+++ cohttp-eio/src/cohttp_eio.ml
@@ -1,3 +1,4 @@
 module Body = Body
+module Chunked_body = Chunked_body
 module Server = Server
 module Client = Client