+++ CHANGES.md
@@ -8,6 +8,7 @@
 - cohttp-eio: further improve Cohttp_eio.Client ergonomics (bikallem #?)
 - cohttp-eio: server api improvements (bikallem #962)
 - cohttp-eio: split Rwer module into Buf_read and Buf_write (bikallem #971)
+- cohttp-eio: extend Body with reader/writer semantics, split Chunked_body (bikallem #?)
 
 ## v6.0.0~alpha0 (2022-10-24)
 - cohttp-eio: ensure "Host" header is the first header in http client requests (bikallem #939)